requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
certifi==2025.10.5
cffi==2.0.0
cloudinary==1.44.1
cryptography==46.0.2
deprecation==2.1.0
dj-database-url==3.0.1
Django==5.2.7
django-cloudinary-storage==0.3.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
multidict==6.7.0
packaging==25.0
pillow==11.3.0
propcache==0.4.1
psycopg==3.2.10
psycopg-binary==3.2.10
pycparser==2.23
pydantic==2.12.0
pydantic_core==2.41.1
PyJWT==2.10.1
python-dotenv==1.1.1
sniffio==1.3.1
sqlparse==0.5.3
StrEnum==0.4.15
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
websockets==15.0.1
yarl==1.22.0
matplotlib==3.8.1
reportlab==4.0.0
drf-spectacular==0.27.0
drf-spectacular-sidecar==2024.1.1
django-cors-headers==4.3.1
gunicorn==21.2.0
sendgrid==6.11.0